+++ api/test/tests/e2e/track_upload/test_upload_track_with_fingerprinting_and_musicbrainz_lookup.py
@@ -26,7 +26,6 @@ ACOUSTID_LOOKUP_RECORDING_PAYLOAD = {
 
 
 @pytest.mark.e2e
-@pytest.mark.usefixtures("enable_audio_metadata_analysis")
 class TestCase(UploadedTrackTestCase):
     """
     E2E test for complete track upload workflow with audio fingerprinting and MusicBrainz lookup.