+++ api/test/tests/integration/uploaded_track/input/method/post/attributes/file/musicbrainz_recording/test_musicbrainz_id.py
@@ -8,7 +8,6 @@ from api.test.utils.uploaded_track.UploadedTrackTestFilename import UploadedTrac
 from api.test.tests.integration.uploaded_track.UploadedTrackTestCase import UploadedTrackTestCase
 
 
-@pytest.mark.usefixtures("enable_audio_metadata_analysis")
 class TestCase(UploadedTrackTestCase):
 
     def test_drown_7m21_mp3_with_mocked_lookup_then_ok(self):